requirements.txt
Flask>=2.2.3
openai==1.57.0
python-dotenv==1.0.1